+++ backend/config/basicAuthData.json
@@ -1,0 +1,4 @@
+﻿{
+  "guest": "MatsuiLab",
+  "fukuyori": "password"
+}
