+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "node server.js",
+    "start": "node backend/index.js",
     "dev": "webpack --mode development",
     "watch": "webpack -w",
     "build": "webpack --mode production"
@@ -27,12 +27,15 @@
     "date-fns-timezone": "^0.1.4",
     "debug": "^4.3.4",
     "express": "^4.17.3",
+    "express-session": "^1.17.3",
     "fabric": "^3.6.3",
     "http-errors": "^2.0.0",
     "moment": "^2.29.3",
     "morgan": "^1.10.0",
     "mysql": "^2.18.1",
     "npm": "^8.5.4",
+    "passport": "^0.6.0",
+    "passport-local": "^1.0.0",
     "react": "^17.0.2",
     "react-contextmenu": "^2.14.0",
     "react-dom": "^17.0.2",
